pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "codeatlas-mcp"
version = "2.6.0"
description = "MCP server for semantic code analysis, repository mapping, and large codebase navigation"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    {name = "NeuroGhostDev"}
]
keywords = [
    "mcp",
    "code-analysis",
    "semantic-search",
    "ai",
    "llm",
    "tree-sitter",
    "faiss",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "fastmcp>=0.1.0",
    "tree-sitter>=0.25.0",
    "watchdog>=4.0.0",
    "faiss-cpu>=1.7.0",
    "sentence-transformers>=2.0.0",
    "numpy>=1.24.0",
    "grep-ast>=0.9.0",
    "tree-sitter-language-pack>=0.13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.urls]
Homepage = "https://github.com/NeuroGhostDev/codeatlas-mcp"
Repository = "https://github.com/NeuroGhostDev/codeatlas-mcp.git"
Issues = "https://github.com/NeuroGhostDev/codeatlas-mcp/issues"

[project.scripts]
codeatlas-mcp = "server:main"

[tool.setuptools.packages.find]
where = ["."]
exclude = ["tests*", "test*"]

[tool.black]
line-length = 100
target-version = ['py312']

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
